+++ cpp/include/cudf/copying.hpp
@@ -22,6 +22,15 @@
 namespace cudf {
 namespace experimental {
 
+/** ---------------------------------------------------------------------------*
+* @brief Indicates when to allocate a mask, based on an existing mask.
+* ---------------------------------------------------------------------------**/
+enum mask_allocation_policy {
+    NEVER, ///< Do not allocate a null mask, regardless of input
+    RETAIN, ///< Allocate a null mask if the input contains one
+    ALWAYS ///< Allocate a null mask, regardless of input
+};
+
 /*
  * Initializes and returns an empty column of the same type as the `input`.
  *
@@ -29,6 +38,35 @@ namespace experimental {
  * @return std::unique_ptr<column> An empty column of same type as `input`
  */
 std::unique_ptr<column> empty_like(column_view input);
+
+/**
+ * @brief Creates an uninitialized new column of the same size and type as the `input`.
+ * Supports only fixed-width types.
+ *
+ * @param input Immutable view of input column to emulate
+ * @param mask_alloc Optional, Policy for allocating null mask. Defaults to RETAIN.
+ * @param mr Optional, The resource to use for all allocations
+ * @return std::unique_ptr<column> A column with sufficient uninitialized capacity to hold the same number of elements as `input` of the same type as `input.type()`
+ */
+std::unique_ptr<column> allocate_like(column_view input,
+		                      mask_allocation_policy mask_alloc = RETAIN,
+                                      rmm::mr::device_memory_resource *mr =
+				          rmm::mr::get_default_resource());
+
+/**
+ * @brief Creates an uninitialized new column of the specified size and same type as the `input`.
+ * Supports only fixed-width types.
+ *
+ * @param input Immutable view of input column to emulate
+ * @param size The desired number of elements that the new column should have capacity for
+ * @param mask_alloc Optional, Policy for allocating null mask. Defaults to RETAIN.
+ * @param mr Optional, The resource to use for all allocations
+ * @return std::unique_ptr<column> A column with sufficient uninitialized capacity to hold the specified number of elements as `input` of the same type as `input.type()`
+ */
+std::unique_ptr<column> allocate_like(column_view input, size_type size,
+                                      mask_allocation_policy mask_alloc = RETAIN,
+                                      rmm::mr::device_memory_resource *mr =
+				          rmm::mr::get_default_resource());
 
 /**
  * @brief Creates a table of empty columns with the same types as the `input_table`